Adição da interface das telas relacionadas ao pagamento, gerenciamento de carros e cartoes

+++ app/src/main/java/com/example/myapplication/model/Cartao.kt
@@ -4,7 +4,7 @@ data class Cartao(
     val id: Int? = null, // O ID pode ser nulo ao criar, o backend irá gerá-lo
     val banco: String,
     val nome: String,
-    val numero: Int,
+    val numero: Long,
     val validade: Int,
     val cvc: Int,
     val userId: Int // Chave estrangeira para associar ao usuário
